Separated FX and Compose UI implementations in different packages (modules TODO). Added flag to start compose UI implementation
