+++ src/main/kotlin/ui/Application.kt
@@ -1,28 +1,12 @@
 package ui
 
-import database.DbHelper
-import javafx.stage.Stage
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import tornadofx.App
-import tornadofx.launch
-import kotlin.system.exitProcess
-
-class Application:  App(ui.main.MainView::class) {
-
-    override fun start(stage: Stage) {
-        super.start(stage)
-
-        stage.isMaximized = true
-
-        stage.setOnCloseRequest {
-            exitProcess(0)
-        }
-    }
-}
+import ui.compose.ComposeApplication
+import ui.fx.FxApplication
 
 fun main(args: Array<String> = emptyArray()) {
-    CoroutineScope(Dispatchers.IO).launch { DbHelper.init() }
-    launch<Application>(args)
+    if ("--compose" in args) {
+        ComposeApplication.main(args)
+    } else {
+        FxApplication.main(args)
+    }
 }
