+++ src/main/kotlin/ui/compose/shared/components/FileChooser.kt
@@ -3,11 +3,8 @@ package ui.compose.shared.components
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.window.WindowScope
 import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.javafx.JavaFx
 import kotlinx.coroutines.launch
 import java.io.File
 import javax.swing.JFileChooser
